make editor component for reuse
